+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog of ng-walkthrough
 
+## V0.10.2 (2024-06-25)
+
+### New feature
+
+-   positioning management when no target
+-   add `top-screen-center` position for `verticalAlignContent` when no target
+
 ## V0.10.1 (2024-06-24)
 
 ### Correction